remove read only pairs from all kvp edits
